test: split the own-key probe test into its own module for the 2000-line cap
object/tests.rs reached 2088 lines, over check_file_size.sh's cap. Extracts
the new #9180 test into object/own_key_probe_tests.rs. Adds the changelog
fragment.
